refactor: Product 클래스의 price 자료형 변경
Product 클래스의 price 자료형 변경 및 관련 DTO 클래스 수정했습니다.

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/GetProductResponse.java
@@ -10,7 +10,7 @@ public record GetProductResponse(
         Long id,
         String name,
         String category,
-        Long price,
+        int price,
         int quantity,
         ProductStatus status,
         LocalDateTime createdAt,
